Fix interval stacking and print ties in chord
Claudio Sacco alerted me to the fact that chord intervals are
transcribed wrongly.  This is an oversight from the time when I
hadnt understood them correctly yet.  Intervals apply to the
first note and then to the last interval, not always to the first note.

Also, print tied notes in chords.