A small throw-away function for Emacs to reduce <divisions>
This is useful to make Dolet for Sibelius exported MusicXML files
(which uses 768 ticks per quarter note by default) more readable.

However, it is not complete and therefore not really useful generically
since it might trash MusicXML files that either use more than
one <divisions> setting or a duration value in an attribute of
an element.  So use with care.