Rewrite handling of fractional values
Replaces the sort of clumsy previous attempt of handling fractions in music
with AbstractFraction, an abstract class that implements all the mathematical
operations based on abstract numerator and denominator methods.
Furthermore, define PowerOfTwo, which extends AbstractFraction,
and replace AugmentedFraction with AugmentedPowerOfTwo which extends
PowerOfTwo.

This commit touches a lot of code but should not change functionality.